Decouple integration tests from microk8s
The integration test needs to use `kubectl` but because tox doesn't
automatically pass through `$HOME` var, it wasn't able to find the kube
config in the default location. Since it was running with microk8s, the
workaround was to explicitly use `microk8s kubectl` but that couples the
test to running in microk8s and prevents it from being run against a
different K8s provider.

This change provides a better fix by explicitly setting and passing
through the `$KUBECONFIG` var instead. It also adds a couple of missing
checks that were causing some commands to fail silently.