fix ingress RelationBrokenEvent handling
Also adds unit tests for the event.  Integration tests are not included but should be added in future to ensure VirtualServices are generated/removed as appropriate

Removes handling the RelationDepartedEvent as (I think) the appropriate event to handle is RelationBrokenEvent.  RelationDeparted happens when any unit on the other side of the relation leaves, whereas RelationBroken happens when all units have left and the relation is being removed.  I think we only care about the latter.