Make cleanup errors in ingress test non-fatal
For some reason, the "remove offer" step of the cleanup creates a new
connection to the controller which tends to fail more often than not in
CI, making the tests much more flakey. Since the CI is already being
done in an ephemeral environment, it's better to just log the issue and
ignore it, since it doesn't impact the test itself.