istio-gateway/src/charm.py: replaces on.install with on.start event
Checking certain conditions, like the existance of a relation, should be
done on StartEvent instead of on InstallEvent, as we can allow the charm
to setup and other events to happen, such as relation-created.
We can ensure that if any check we have in this event handler is failing,
it is due to an omission and not because we didn't allow the charm to
actually meet those conditions.