istio-pilot/src/charm.py: implements a reconcile resources method
In previous versions of this file we are removing all resources of
a kind on different events and then applying them again. This could
raise issues if any related charm is using such resource at the moment
of removal.
Using a reconcile helper method allows us to keep alive the existing
resources and only remove those that are not backed by a relation.
In this commit this method is implemented and also used for the
handle_ingress event handler as a fix for #50.

Fixes #50