istio-pilot/src/resource_handler.py: implements a reconcile resources method
In previous versions of this file we are removing all resources of
a kind and then applying them again w/o caring about their origin.
This could raise issues if any related charm is using such resource at the moment
of removal.
Using a reconcile helper method allows us to keep alive the existing
resources and only remove those that are not backed by a relation.
In this commit this method is implemented and also used for the
handle_ingress event handler as a fix for #50.

Fixes #50

Co-authored-by: Daniela Plascencia <daniela.plascencia@canonical.com>
Co-authored-by: Andrew Scribner <ca.scribner+1@gmail.com>