istio-pilot: moves mocked_list out of conftest.py
The way Client.list is mocked in conftest.py does not apply to all
test cases as it is directly building the resource's metadata
and other parameters, which might not be useful for tests where
the resources metadata is provided by the test case itself.
Extracting this mock and leaving it inside test_charm.py prevents
other test modules to use it as a fixture.