test-requirements.txt: pins cryptography==36.0.2 due to deps issues
A recent change in cryptography, a package installed as dependency during
integration testing, causes a Warning in another dependency (paramiko),
that prevents building/packaging the charm to be tested.
From paramiko/paramiko#2038 [1], it seems cryptography's latest version is the
one causing the error due to some deprecated package.
Pinnin cryptography to version 36.0.2 prevents the deprecation warning.

[1] https://github.com/paramiko/paramiko/issues/2038