istio-pilot: send data whenever we handle ingress
Relation data is only sent once on istio-pilot.relation_changed,
which works fine when relating just one charm, but when multiple
charms are related, a series of unmet conditions make the charm to
cycle and never send the data.
This change ensures sending the data whenever handle_ingress is called.