refactor: use log.debug and log.error on certain exceptions in istio-pilot charm code
The previous implementation logged all exceptions, which made
logs very noisy and led to confusion. Changing that to use log.debug
when the exception is of TypeError or ApiError, and use log.error when
an unexpected error is caught. This commit also changes the event.defer()
call so that it only happens on expected errors.