fix: bind handle_default_gateway to ingress_relation_created
The handle_default_gateway method is not run on config_changed at deployment time,
this is an unexpected behaviour that happens when istio-pilot is related to
multiple other charms in a bundle. Although config_changed is guaranteed to
run after install, that is not the case under the described conditions, causing
the default gateway to not be created without human intervention.
To avoid this situation, and given that istio-pilot is usually deployed in a bundle
and related to multiple charms, the handle_default_gateway method can be also bound to
["ingress"].relation_created, which is also guaranteed to run after install.

Fixes: canonical/istio-operators#94