feat: add option to use `NodePort` for istio-gateway (#176)
This enables the istio-gateway charm to provide a gateway service as a `NodePort`, and refactors istio-pilot to properly handle `NodePort` services.  Also added here are unit tests to assert these new features.

In implementing this, some existing tests were found to pass when they shouldn't.  This was highlighted when refactoring the istio-pilot pulling of the Gateway IP/hostport.  These have been reworked here too.

Not added here are integration tests for the `NodePort` option.  This has been manually tested, and integration tests will be added before the next main release.