fix: support ingress schema v1 in reconcile_ingress
Add a default namespace for any ingress relation.  If the Service's Namespace is not specified, assume it is the same as istio-pilot's.

The ingress v2 schema added the namespace field to support cross-namespace ingresses.  Previously, the ingress v1 schema did not include the namespace and istio-pilot always assumed everything was in istio-pilot's namespace.