fix: handle ingress relation-broken if event.app==None
Fixes [this issue](https://github.com/canonical/istio-operators/issues/189).

While handling relation-broken events for ingress, we see either:
* event.app=the_departing_app
* event.app=None

The previous charm logic did not cover the second case, leading to an uncaught KeyError.

We aren't sure what controls which of the situations we encounter, but we add error handling here to at least log the event and avoid uncaught exceptions.