feat: enable certificates relation on istio-pilot
Enabling the certificates relation enables the integration with tls-certificates-operator
and other charms that provide CA certificates. These CA certificates are used by istio-pilot
to configure TLS on the ingress Gateway resource.
In the past this was done exclusively via charm configuration, which is now removed in favour
of CA certificates provided through the certificates relation.