fix: shorten the summary as a workaround for canonical/charmcraft#1568
In recent versions of charmcraft, the summary line must not exceed 78 characters. This change
has been introduced in v2.5, which is heavily used by CKF charms. Since pinning the charmcraft
version to an earlier release won't scale correctly, it has been decided to move on with the
workaround, which alings more to what charmcraft will expect in the future.