build: bump juju, ops versions in ci and requirements files
Bumping juju and ops packages to use them in newer versions of the charms,
plus testing them in a CI with a more recent juju version.

This commit also skips some test cases that will be removed in a follow
up commit introduced by canonical/istio-operators#401.

Part of canonical/bundle-kubeflow#859
Part of #398

Signed-off-by: Daniela Plascencia <daniela.plascencia@canonical.com>