chore: use assumes juju 3.1 instead of 3.5
We must ensure we do not break the deployment of existing users
with a hard requirement on juju 3.5. Since in our documentation the least supported
juju version is 3.1, we should reflect that also in these charms.
The only scenario where people are required to bump juju to 3.5 is when they want to
set the TLS ingress gateway with secrets, but this is documented in a public guide.