tests: point all the integration charm deps to their 1.7 versions (#439)
The integration tests were deploying versions of dependency charms that do
not correspond to the current release of the istio-operators. This commit brings
all the correct versions and configure charms accordingly to ensure the integration
tests execute in the same environment as this version of the istio-operators were
design for.