tests: remove public-url config and add dex-oidc-config relation
The public-url configuration option has been replaced by
a relation between dex-auth and oidc-gatekeeper. Because of that,
configuring this option is not required anymore, and in fact if kept,
can cause errors as it won't be recognised by Juju anymore.
This commit removes the extra configuration from the integration tests.

Part of #487