chore: remove deprecation warning for the TLS secret configuration
Originally, the TLS configuration that's done by passing a juju secret to istio-pilot
was expected to last only a few versions of this charm, but as exposed in #536, this
feature is needed until there is a solution that fits out users/customer needs better.
Because of that, this commit removes the comments and warnings about deprecating this
feature.