fix(tls): use cert_subject in the sans and cert_subject fields (#614)
In a previous version of the istio-pilot charm, the sans argument of
the CertHandler was not set, meaning that the library will always use the FQDN
of the Istio Ingress Service as the value for the SANS in the CSR.
While this is an implementation detail of the CertHandler library, it was impossible
for users to set it correctly using their desired DNS names.
This commit ensures the SANS is pointing to the csr-domain-name (if configured)
or the service's IP address.

Fixes #557